android port of libbirdfont

+++ scripts/build.py
@@ -37,7 +37,7 @@ def libbirdfont(prefix, cc, cflags, ldflags, valac, valaflags, library):
 		-H build/libbirdfont/birdfont.h \
 		libbirdfont/*.vala \
 		libbirdfont/OpenFontFormat/*.vala \
-		--pkg gee-1.0 \
+		--pkg gee-0.8 \
 		--pkg libxml-2.0 \
 		--pkg gio-2.0 \
 		--pkg cairo \
@@ -46,41 +46,44 @@ def libbirdfont(prefix, cc, cflags, ldflags, valac, valaflags, library):
 	#copy c sources 
 	run("cp libbirdfont/OpenFontFormat/*.c build/libbirdfont/")
 
-	run(cc + " " + cflags + """ \
-		-c build/libbirdfont/*.c \
-		-fPIC \
-		-D 'GETTEXT_PACKAGE="birdfont"' \
-		$(pkg-config --cflags gee-1.0) \
-		$(pkg-config --cflags libxml-2.0) \
-		$(pkg-config --cflags gio-2.0) \
-		$(pkg-config --cflags cairo) \
-		$(pkg-config --cflags glib-2.0) \
-		$(pkg-config --cflags gdk-pixbuf-2.0) \
-		-I ./build/mac/birdfont""")
-	run("mv ./*.o build/libbirdfont/ ")
+	if cc == "":
+		print ("Skipping compilation");
+	else:
+		run(cc + " " + cflags + """ \
+			-c build/libbirdfont/*.c \
+			-fPIC \
+			-D 'GETTEXT_PACKAGE="birdfont"' \
+			$(pkg-config --cflags gee-0.8) \
+			$(pkg-config --cflags libxml-2.0) \
+			$(pkg-config --cflags gio-2.0) \
+			$(pkg-config --cflags cairo) \
+			$(pkg-config --cflags glib-2.0) \
+			$(pkg-config --cflags gdk-pixbuf-2.0) \
+			-I ./build/mac/birdfont""")
+		run("mv ./*.o build/libbirdfont/ ")
 
-	run(cc + " " + ldflags + """ \
-		-shared \
-		build/libbirdfont/*.o \
-		$(freetype-config --libs) \
-		$(pkg-config --libs gee-1.0) \
-		$(pkg-config --libs libxml-2.0) \
-		$(pkg-config --libs gio-2.0) \
-		$(pkg-config --libs cairo) \
-		$(pkg-config --libs glib-2.0) \
-		$(pkg-config --libs gdk-pixbuf-2.0) \
-		-o """ + library)
-	run("mv " + library + " build/bin/")
-	
-	if os.path.exists("build/bin/libbirdfont.so"):
-		run ("cd build/bin && unlink libbirdfont.so")
-
-	# create link to the versioned library
-	if library.find ('.so') > -1:
-		run ("""cd build/bin && ln -s """ + library + " libbirdfont.so")
-	elif library.find ('.dylib') > -1:
-		run ("""cd build/bin && ln -s """ + library + " libbirdfont.dylib")
+		run(cc + " " + ldflags + """ \
+			-shared \
+			build/libbirdfont/*.o \
+			$(freetype-config --libs) \
+			$(pkg-config --libs gee-0.8) \
+			$(pkg-config --libs libxml-2.0) \
+			$(pkg-config --libs gio-2.0) \
+			$(pkg-config --libs cairo) \
+			$(pkg-config --libs glib-2.0) \
+			$(pkg-config --libs gdk-pixbuf-2.0) \
+			-o """ + library)
+		run("mv " + library + " build/bin/")
 		
+		if os.path.exists("build/bin/libbirdfont.so"):
+			run ("cd build/bin && unlink libbirdfont.so")
+
+		# create link to the versioned library
+		if library.find ('.so') > -1:
+			run ("""cd build/bin && ln -s """ + library + " libbirdfont.so")
+		elif library.find ('.dylib') > -1:
+			run ("""cd build/bin && ln -s """ + library + " libbirdfont.dylib")
+			
 	
 def birdfont_export(prefix, cc, cflags, ldflags, valac, valaflags, library):
 	# birdfont-export
@@ -92,7 +95,7 @@ def birdfont_export(prefix, cc, cflags, ldflags, valac, valaflags, library):
 		--enable-experimental \
 		--define=MAC birdfont-export/* \
 		--vapidir=./ \
-		--pkg gee-1.0 \
+		--pkg gee-0.8 \
 		--pkg libxml-2.0 \
 		--pkg gio-2.0  \
 		--pkg cairo \
@@ -104,7 +107,7 @@ def birdfont_export(prefix, cc, cflags, ldflags, valac, valaflags, library):
 	run(cc + " " + cflags + """ \
 		-c ./build/libbirdfont/birdfont.h build/birdfont-export/*.c \
 		-D 'GETTEXT_PACKAGE="birdfont"' \
-		$(pkg-config --cflags gee-1.0) \
+		$(pkg-config --cflags gee-0.8) \
 		$(pkg-config --cflags libxml-2.0) \
 		$(pkg-config --cflags gio-2.0) \
 		$(pkg-config --cflags cairo) \
@@ -116,7 +119,7 @@ def birdfont_export(prefix, cc, cflags, ldflags, valac, valaflags, library):
 	run(cc + " " + ldflags + " \
 		build/birdfont-export/*.o \
 		-Lbuild/bin/ -lbirdfont \
-		$(pkg-config --libs gee-1.0) \
+		$(pkg-config --libs gee-0.8) \
 		$(pkg-config --libs libxml-2.0) \
 		$(pkg-config --libs gio-2.0) \
 		$(pkg-config --libs cairo) \
@@ -137,7 +140,7 @@ def birdfont_gtk(prefix, cc, cflags, ldflags, valac, valaflags, library):
 		--enable-experimental-non-null \
 		--enable-experimental \
 		--vapidir=./ \
-		--pkg gee-1.0 \
+		--pkg gee-0.8 \
 		--pkg libxml-2.0 \
 		--pkg gio-2.0  \
 		--pkg cairo \
@@ -151,7 +154,7 @@ def birdfont_gtk(prefix, cc, cflags, ldflags, valac, valaflags, library):
 	run(cc + " " + cflags + """\
 		-c ./build/libbirdfont/birdfont.h build/birdfont/*.c \
 		-D 'GETTEXT_PACKAGE="birdfont"' \
-		$(pkg-config --cflags gee-1.0) \
+		$(pkg-config --cflags gee-0.8) \
 		$(pkg-config --cflags libxml-2.0) \
 		$(pkg-config --cflags gio-2.0) \
 		$(pkg-config --cflags cairo) \
@@ -164,7 +167,7 @@ def birdfont_gtk(prefix, cc, cflags, ldflags, valac, valaflags, library):
 	run(cc + " " + ldflags + """ \
 		build/birdfont/*.o \
 		-Lbuild/bin/ -lbirdfont \
-		$(pkg-config --libs gee-1.0) \
+		$(pkg-config --libs gee-0.8) \
 		$(pkg-config --libs libxml-2.0) \
 		$(pkg-config --libs gio-2.0) \
 		$(pkg-config --libs cairo) \
